+++ src/ansys/mechanical/core/run.py
@@ -171,6 +171,11 @@ def cli(
     if input_script:
         args.append("-script")
         args.append(input_script)
+        warnings.warn(
+            "Please ensure ExtAPI.Application.Close() is at the end of your script. "
+            "Without this command, Batch mode will not terminate.",
+            stacklevel=2,
+        )
 
         if version < 241:
             warnings.warn(