+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This document follows the conventions laid out in [Keep a CHANGELOG](https://kee
 
 ### Added
 
+-   Added warning for ansys-mechanical when provided an input script (#319)
+
 ### Changed
 
 ### Fixed